pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-forecast"
version = "0.1.0"
description = "BTC 1h ahead price forecast from 1m OHLCV using LightGBM"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "YM Quant Candidate" }]
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "lightgbm>=4.3.0",
  "scikit-learn>=1.4.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0",
  "typer>=0.12.3",
  "loguru>=0.7.2",
  "matplotlib>=3.8.0",
  "pyarrow>=15.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0"
]

[project.scripts]
btc-predict = "btc_forecast.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
